src/main/java/lgwarlick/services/IngredientServiceImpl.java
package lgwarlick.services;

import lgwarlick.commands.IngredientCommand;
import lgwarlick.converters.IngredientToIngredientCommand;
import lgwarlick.domain.Recipe;
import lgwarlick.repositories.RecipeRepository;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Slf4j
@Service
public class IngredientServiceImpl implements IngredientService{

    private final IngredientToIngredientCommand ingredientToIngredientCommand;
    private final RecipeRepository recipeRepository;

    public IngredientServiceImpl(IngredientToIngredientCommand ingredientToIngredientCommand, RecipeRepository recipeRepository) {
        this.ingredientToIngredientCommand = ingredientToIngredientCommand;
        this.recipeRepository = recipeRepository;
    }

    @Override
    public IngredientCommand findByRecipeIdAndIngredientId(long recipeId, long ingredientId) {


        Optional<Recipe> recipeOptional = recipeRepository.findById(recipeId);

        if (!recipeOptional.isPresent()){
            log.error("Recipe id not found for id: " + recipeId);
        }

        Recipe recipe = recipeOptional.get();

        //stream to search for the id that was passed, then convert to a ingredientCommand
        Optional<IngredientCommand> ingredientCommandOptional = recipe.getIngredients().stream()
                .filter(ingredient -> ingredient.getId().equals(ingredientId))
                .map(ingredient -> ingredientToIngredientCommand.convert(ingredient)).findFirst();

        if (!ingredientCommandOptional.isPresent()){
            log.error("Ingredient id not found for: " + ingredientId);
        }

        return ingredientCommandOptional.get();
    }
}
